Add userId check in dashboard widget and net worth items

+++ apps/backend/src/main/dashboard/logic/delete-dashboard-widget-service.ts
@@ -1,16 +1,28 @@
-import { Injectable } from "@nestjs/common";
-import { PrismaService } from "../../../prisma/prisma.service";
-
+import { Injectable, NotFoundException } from '@nestjs/common';
+import { PrismaService } from '../../../prisma/prisma.service';
 
 @Injectable()
 export class DeleteDashboardWidget {
-    constructor(private readonly prisma: PrismaService) {}
+  constructor(private readonly prisma: PrismaService) {}
 
-    async delete(widgetId: string) {
-        return this.prisma.dashboardWidget.delete({
-            where: {
-                id: widgetId
-            }
-        })
+  async delete(userId: string, widgetId: string) {
+    const widget = await this.prisma.dashboardWidget.findFirst({
+      where: {
+        id: widgetId,
+        dashboard: {
+          userId,
+        },
+      },
+    });
+
+    if (!widget) {
+      throw new NotFoundException('Widget not found');
     }
+
+    return this.prisma.dashboardWidget.delete({
+      where: {
+        id: widgetId,
+      },
+    });
+  }
 }
