+++ apps/backend/src/main/net-worth/logic/delete-net-worth-item-service.ts
@@ -1,13 +1,26 @@
-import { Injectable } from "@nestjs/common";    
-import { PrismaService } from "../../../prisma/prisma.service";
+import { Injectable, NotFoundException } from '@nestjs/common';
+import { PrismaService } from '../../../prisma/prisma.service';
 
 @Injectable()
 export class DeleteNetWorthItemService {
-    constructor(private readonly prisma: PrismaService) {}
+  constructor(private readonly prisma: PrismaService) {}
 
-    async execute( itemId: string) {
-        return this.prisma.netWorthItem.delete({
-            where: { id: itemId },   
-        })
+  async execute(userId: string, itemId: string) {
+    const item = await this.prisma.netWorthItem.findFirst({
+      where: {
+        id: itemId,
+        snapshot: {
+          userId,
+        },
+      },
+    });
+
+    if (!item) {
+      throw new NotFoundException('Item not found');
     }
-}   
+
+    return this.prisma.netWorthItem.delete({
+      where: { id: itemId },
+    });
+  }
+}
